+++ src/importlinter/contracts/forbidden.py
@@ -14,8 +14,8 @@ class ForbiddenContract(Contract):
 
         - source_modules:    A list of Modules that should not import the forbidden modules.
         - forbidden_modules: A list of Modules that should not be imported by the source modules.
-        - ignore_imports:    A list of DirectImports. These imports will be ignored: if the import
-                             would cause a contract to be broken, adding it to the list will cause
+        - ignore_imports:    A set of DirectImports. These imports will be ignored: if the import
+                             would cause a contract to be broken, adding it to the set will cause
                              the contract be kept instead. (Optional.)
     """
 
@@ -23,7 +23,7 @@ class ForbiddenContract(Contract):
 
     source_modules = fields.ListField(subfield=fields.ModuleField())
     forbidden_modules = fields.ListField(subfield=fields.ModuleField())
-    ignore_imports = fields.ListField(subfield=fields.DirectImportField(), required=False)
+    ignore_imports = fields.SetField(subfield=fields.DirectImportField(), required=False)
 
     def check(self, graph: ImportGraph) -> ContractCheck:
         is_kept = True