+++ src/importlinter/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.2"
+__version__ = "1.2.1"
 
 from .application import output  # noqa
 from .domain import fields  # noqa