+++ CHANGELOG.rst
@@ -87,3 +87,4 @@ latest
 ------
 
 * Upgrade Grimp to 1.2.2.
+* Drop duplicated values of `ignore_imports` options