+++ CHANGELOG.rst
@@ -87,3 +87,5 @@ latest
 ------
 
 * Upgrade Grimp to 1.2.2.
+* Add SetField
+* Use a SetField for ignore_imports options